Update PR according to review
