+++ source/components/organisms/CheckboxList/CheckboxList.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
-import styled from "styled-components/native";
 
-import { Checkbox, Text } from "app/components/atoms";
+import { Checkbox } from "../../atoms";
 import { ComplementaryColor } from "../../../styles/themeHelpers";
+import { BoxTextWrapper, CheckboxFieldText } from "./CheckboxList.styled";
 
 interface CheckboxListProps {
   choices: { displayText: string; tags?: string[]; id: string }[];
@@ -10,38 +10,8 @@ interface CheckboxListProps {
   onBlur?: (values: Record<string, boolean>) => void;
   value?: Record<string, boolean>;
   colorSchema?: ComplementaryColor;
+  size?: "small" | "medium" | "large";
 }
-
-const BoxTextWrapper = styled.View`
-  flex: auto;
-  flex-direction: row;
-  align-items: flex-start;
-`;
-
-const sizes = {
-  small: {
-    padding: 0.25,
-    margin: 4,
-    marginTop: 0,
-    fontSize: 12,
-  },
-  medium: {
-    padding: 0.5,
-    margin: 4,
-    fontSize: 15,
-  },
-  large: {
-    padding: 1,
-    margin: 5,
-    fontSize: 18,
-  },
-};
-
-const CheckboxFieldText = styled(Text)<{ size: "small" | "medium" | "large" }>`
-  margin-left: ${(props) => props.theme.sizes[1]}px;
-  margin-right: ${(props) => props.theme.sizes[1]}px;
-  font-size: ${(props) => sizes[props.size].fontSize}px;
-`;
 
 const CheckboxList = (props: CheckboxListProps): JSX.Element => {
   const {
@@ -50,6 +20,7 @@ const CheckboxList = (props: CheckboxListProps): JSX.Element => {
     onBlur,
     value: answerValues,
     colorSchema = "red",
+    size = "small",
   } = props;
 
   const [values, setValues] = useState<Record<string, boolean>>({});
@@ -83,12 +54,12 @@ const CheckboxList = (props: CheckboxListProps): JSX.Element => {
         <BoxTextWrapper key={choice.id}>
           <Checkbox
             colorSchema={colorSchema}
-            size="small"
+            size={size}
             onChange={() => update(choice.id)}
-            checked={values[choice.id]}
+            checked={values[choice.id] ?? false}
             // {...other}
           />
-          <CheckboxFieldText color={colorSchema} size="small">
+          <CheckboxFieldText size={size}>
             {choice.displayText}
           </CheckboxFieldText>
         </BoxTextWrapper>
